+++ tfjs-backend-cpu/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "@tensorflow/tfjs-backend-cpu",
+  "version": "1.5.2",
+  "description": "Hardware-accelerated JavaScript library for machine intelligence",
+  "private": true,
+  "engines": {
+    "yarn": ">= 1.3.2"
+  },
+  "license": "Apache-2.0",
+  "devDependencies": {
+    "tslint": "~5.11.0",
+    "tslint-no-circular-imports": "~0.5.0",
+    "typescript": "3.5.3"
+  },
+  "dependencies": {
+    "@types/seedrandom": "2.4.27",
+    "seedrandom": "2.4.3"
+  }
+}
