+++ tfjs-backend-cpu/src/backend_cpu.ts
@@ -17,35 +17,35 @@
 
 import * as seedrandom from 'seedrandom';
 
-import {BackendTimingInfo, DataStorage, EPSILON_FLOAT32, KernelBackend} from '../../../backends/backend';
-import * as backend_util from '../../../backends/backend_util';
-import * as complex_util from '../../../backends/complex_util';
-import {nonMaxSuppressionV3Impl} from '../../../backends/non_max_suppression_impl';
-import {split} from '../../../backends/split_shared';
-import {tile} from '../../../backends/tile_impl';
-import {topkImpl} from '../../../backends/topk_impl';
-import {whereImpl} from '../../../backends/where_impl';
-import {ENGINE} from '../../../engine';
-import {env} from '../../../environment';
-import {warn} from '../../../log';
-import * as array_ops_util from '../../../ops/array_ops_util';
-import * as axis_util from '../../../ops/axis_util';
-import * as broadcast_util from '../../../ops/broadcast_util';
-import {complex, imag, real} from '../../../ops/complex_ops';
-import * as concat_util from '../../../ops/concat_util';
-import {Conv2DInfo, Conv3DInfo} from '../../../ops/conv_util';
-import * as erf_util from '../../../ops/erf_util';
-import {Activation, FusedBatchMatMulConfig, FusedConv2DConfig} from '../../../ops/fused_util';
-import * as gather_nd_util from '../../../ops/gather_nd_util';
-import * as ops from '../../../ops/ops';
-import {buffer, scalar, tensor, tensor4d} from '../../../ops/ops';
-import * as scatter_nd_util from '../../../ops/scatter_nd_util';
-import * as selu_util from '../../../ops/selu_util';
-import {computeFlatOffset, computeOutShape, isSliceContinous} from '../../../ops/slice_util';
-import {DataId, Scalar, Tensor, Tensor1D, Tensor2D, Tensor3D, Tensor4D, Tensor5D, TensorBuffer} from '../../../tensor';
-import {BackendValues, DataType, DataValues, NumericDataType, Rank, ShapeMap, TypedArray, upcastType} from '../../../types';
-import * as util from '../../../util';
-import {getArrayFromDType, inferDtype, now, sizeFromShape} from '../../../util';
+import {BackendTimingInfo, DataStorage, EPSILON_FLOAT32, KernelBackend} from '../../tfjs-core/src/backends/backend';
+import * as backend_util from '../../tfjs-core/src/backends/backend_util';
+import * as complex_util from '../../tfjs-core/src/backends/complex_util';
+import {nonMaxSuppressionV3Impl} from '../../tfjs-core/src/backends/non_max_suppression_impl';
+import {split} from '../../tfjs-core/src/backends/split_shared';
+import {tile} from '../../tfjs-core/src/backends/tile_impl';
+import {topkImpl} from '../../tfjs-core/src/backends/topk_impl';
+import {whereImpl} from '../../tfjs-core/src/backends/where_impl';
+import {ENGINE} from '../../tfjs-core/src/engine';
+import {env} from '../../tfjs-core/src/environment';
+import {warn} from '../../tfjs-core/src/log';
+import * as array_ops_util from '../../tfjs-core/src/ops/array_ops_util';
+import * as axis_util from '../../tfjs-core/src/ops/axis_util';
+import * as broadcast_util from '../../tfjs-core/src/ops/broadcast_util';
+import {complex, imag, real} from '../../tfjs-core/src/ops/complex_ops';
+import * as concat_util from '../../tfjs-core/src/ops/concat_util';
+import {Conv2DInfo, Conv3DInfo} from '../../tfjs-core/src/ops/conv_util';
+import * as erf_util from '../../tfjs-core/src/ops/erf_util';
+import {Activation, FusedBatchMatMulConfig, FusedConv2DConfig} from '../../tfjs-core/src/ops/fused_util';
+import * as gather_nd_util from '../../tfjs-core/src/ops/gather_nd_util';
+import * as ops from '../../tfjs-core/src/ops/ops';
+import {buffer, scalar, tensor, tensor4d} from '../../tfjs-core/src/ops/ops';
+import * as scatter_nd_util from '../../tfjs-core/src/ops/scatter_nd_util';
+import * as selu_util from '../../tfjs-core/src/ops/selu_util';
+import {computeFlatOffset, computeOutShape, isSliceContinous} from '../../tfjs-core/src/ops/slice_util';
+import {DataId, Scalar, Tensor, Tensor1D, Tensor2D, Tensor3D, Tensor4D, Tensor5D, TensorBuffer} from '../../tfjs-core/src/tensor';
+import {BackendValues, DataType, DataValues, NumericDataType, Rank, ShapeMap, TypedArray, upcastType} from '../../tfjs-core/src/types';
+import * as util from '../../tfjs-core/src/util';
+import {getArrayFromDType, inferDtype, now, sizeFromShape} from '../../tfjs-core/src/util';
 
 import {assertNotComplex} from './utils/tensor_utils';
