+++ tfjs-backend-cpu/src/shared.ts
